[14.0][IMP] acount_invoice_refund_link: add index
